beforeTestingAndDeployment

+++ MScInvoice.UI/wwwroot/js/invoices/items.js
@@ -1,6 +1,11 @@
 ﻿var app = new Vue({
     el: '#app',
     data: {
+        priceIsValid: true,
+        taxIsValid: true,
+        nameIsValid: true,
+        formIsValid: true,
+
         editing: false,
         id: 0,
         loading: false,
@@ -13,6 +18,8 @@
         },
         items: []
     },
+
+
     mounted() {
         this.getItems()
     },
@@ -52,34 +59,78 @@
                 });
         },
         createItem() {
-            this.loading = true;
-            axios.post('/items', this.itemModel)
-                .then(res => {
-                    console.log(res.data);
-                    this.items.push(res.data);
-                })
-                .catch(err => {
-                    console.log(err);
-                })
-                .then(() => {
-                    this.loading = false;
-                    this.editing = false;
-                });
+            this.priceIsValid = typeof this.itemModel.price === 'number'
+                && this.itemModel.price > 0;
+            this.taxIsValid = typeof this.itemModel.tax === 'number'
+                && this.itemModel.tax >= 0
+                && this.itemModel.tax < 100;
+            this.nameIsValid = !!this.itemModel.name
+
+            this.formIsValid = this.priceIsValid
+                && this.taxIsValid
+                && this.nameIsValid;
+            if (this.formIsValid) {
+                this.loading = true;
+                axios.post('/items', this.itemModel)
+                    .then(res => {
+                        console.log(res.data);
+                        this.items.push(res.data);
+                    })
+                    .catch(err => {
+                        this.formIsValid = false;
+                        console.log(err);
+                    })
+                    .then(() => {
+                        this.loading = false;
+                        if (!this.formIsValid) {
+                            this.editing = true;
+                        }
+                        else {
+                            this.editing = false;
+                        }
+                    });
+            }
+            else {
+                console.log("NOT VALID");
+            }
         },
         updateItem() {
-            this.loading = true;
-            axios.put('/items', this.itemModel)
-                .then(res => {
-                    console.log(res.data);
-                    this.items.splice(this.objectIndex, 1, res.data);
-                })
-                .catch(err => {
-                    console.log(err);
-                })
-                .then(() => {
-                    this.loading = false;
-                    this.editing = false;
-                });
+            
+            this.priceIsValid = typeof this.itemModel.price === 'number' && this.itemModel.price > 0;
+            this.taxIsValid = typeof this.itemModel.tax === 'number' && this.itemModel.tax >= 0 && this.itemModel.tax < 100;
+            this.nameIsValid = !!this.itemModel.name
+
+            this.formIsValid = this.priceIsValid && this.taxIsValid && this.nameIsValid
+
+            if (this.formIsValid) {
+
+
+                this.loading = true;
+                axios.put('/items', this.itemModel)
+                    .then(res => {
+                        console.log(res.data);
+                        this.items.splice(this.objectIndex, 1, res.data);
+                    })
+                    .catch(err => {
+                        this.formIsValid = false;
+                        console.log(err);
+                        
+                    })
+                    .then(() => {
+                        console.log("VALID");
+                       
+                        this.loading = false;
+                        if (!this.formIsValid) {
+                            this.editing = true;
+                        }
+                        else {
+                            this.editing = false;
+                        }
+                    });
+            }
+            else {
+                console.log("NOT VALID");
+            }
         },
         editItem(id, index) {
             this.objectIndex = index;
@@ -87,18 +138,21 @@
             this.editing = true;
         },
         deleteItem(id, index) {
-            this.loading = true;
-            axios.delete('/items/' + id)
-                .then(res => {
-                    console.log(res);
-                    this.items.splice(index, 1);
-                })
-                .catch(err => {
-                    console.log(err);
-                })
-                .then(() => {
-                    this.loading = false;
-                });
+            if (confirm("Are you sure you want to delete this item")) {
+                this.loading = true;
+                axios.delete('/items/' + id)
+                    .then(res => {
+                        console.log(res);
+                        this.items.splice(index, 1);
+                    })
+                    .catch(err => {
+                        console.log(err);
+                    })
+                    .then(() => {
+                        this.loading = false;
+                        alert("Item deleted...")
+                    });
+            } 
         },
         newItem() {
             this.editing = true;
@@ -106,6 +160,10 @@
         },
         cancel() {
             this.editing = false;
+            this.priceIsValid = true,
+                this.taxIsValid = true,
+                this.nameIsValid = true,
+            this.formIsValid = true
         }
     },
     computed: {
